santaclaus works

+++ src/classes/_03_12_23/Bridge.java
@@ -61,7 +61,7 @@ class Generator extends Thread {
 
             mutex.unlock();
             try {
-                sleep((long) (Math.random() * 1000));
+                sleep((int) (Math.random() * 1000));
             } catch (InterruptedException e) {
                 throw new RuntimeException(e);
             }
